Make Getting Started example compile
Unlike other code I've seen in the guide, this code didn't compile. It's especially bad that it can be the first experience someone has with actix-web, and then doesn't work. I'm going to run through the docs as a first-time user so I'll try to make sure everything works the way it's supposed to.