Update a method name
The HttpServer::start() was renamed to run() in https://github.com/actix/actix-web/commit/6a0cd2dced8b085865e966e2d40baf1b8ad9d640, but run() does not return the server address.  It looks like there used to be a spawn() method that worked that way, but now the only way to get the server's address seems to be addrs() or addrs_with_scheme().