Clarify use of `Arc` in shared state
**This Commit**
Adjusts the explanation of shared state in the Extractors page.

**Why?**
There was some miscommunication in #219 that resulted in a doc change
being merged without an associated code change. I was about to make
the code change to match the docs and realized that a better example
would show how one could only share _part_ of the application state
using an `Arc`. So, instead of making the Extractor section a
near-duplicate of the Shared Mutable State section, I tried my best to
let them both explain different aspects of state-sharing.