Improve error handler example
**This Commit**
Makes a few changes to the current error handler middleware example:

- renames `render_500` to make it more clear what it does
- changes the handler to return a 500 so the middleware is exercised

**Why?**
I was confused because I was trying to understand this middleware and,
looking at this example, I was wondering under what circumstances the
middleware would trigger. I couldn't tell if it was some magic. I also
couldn't tell what was going on with the extra `head()` route - I
figured that was relevant to the handlers.

After futzing with it a while it seemed to me like we just need to
return an error from the handler that corresponded to the status code
the middleware was looking for so I made that change.